Added 3 js validation functions

+++ page/mixDetail-view.php
@@ -17,29 +17,35 @@
         <h3>Comments</h3>
         <?php
         if ($comments) {
-            for($i = 0; $i < count($comments); $i++) {
-                if($contacts[$i]->getPortrait()){
+            for ($i = 0; $i < count($comments); $i++) {
+                if ($contacts[$i]->getPortrait()) {
                     $portrait = $contacts[$i]->getPortrait();
-                }else{
+                } else {
                     $portrait = $defaultPortrait;
                 }
                 echo '
                     <div class="comment">
                         <div class="commentImageWrap">
-                            <img src="'.$portrait.'">
+                            <img src="' . $portrait . '">
                         </div>
-                        <p class="commentStamp">'.Utils::convertTimestamp($comments[$i]->getUploadDate()).'</p>
-                        <h5>'.$users[$i]->getUsername().'</h5>
+                        <p class="commentStamp">' . Utils::convertTimestamp($comments[$i]->getUploadDate()) . '</p>
+                        <h5>' . $users[$i]->getUsername() . '</h5>
                         <p>' . $comments[$i]->getComment() . '</p>
                     </div>';
             }
         } else {
-            echo '<p>Leave a comment...</p>';
+            if (array_key_exists('username', $_SESSION)) {
+                echo '<p>Leave a comment...</p>';
+            } else {
+                echo '<p>Sign in to leave a comment...</p>';
+            }
         }
-        ?>
-        <form action="index.php?page=mixDetail&mixId=<?php echo $mix->getId(); ?>" method="post">
-            <input type="text" name="comment"/><input type="submit" name="submit" value="Send"/>
-        </form>
+        if (array_key_exists('username', $_SESSION)) {
+            ?>
+            <form action="index.php?page=songDetail&songId=<?php echo $song->getId(); ?>" method="post">
+                <input type="text" name="comment"/><input type="submit" name="submit" value="Send"/>
+            </form>
+<?php } ?>
     </div>
 </div>
